rename: rename test package
