+++ src/test/java/racing/FinalResultTest.java
@@ -1,4 +1,4 @@
-package study;
+package racing;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
