+++ src/test/java/racing/MovementHistoryTest.java
@@ -1,4 +1,4 @@
-package study;
+package racing;
 
 import static org.assertj.core.api.Assertions.assertThat;
