+++ src/test/java/racing/ValidatorTest.java
@@ -1,6 +1,5 @@
-package study;
+package racing;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
